Move caching module to frontend
